refactor: even more methods
